setup architecture
