+++ app/src/main/java/com/estudos/nuhome/home/data/NuHomeRepository.kt
@@ -1,9 +1,8 @@
 package com.estudos.nuhome.home.data
 
-import com.estudos.nuhome.home.data.model.response.HomeUserDetailsResponse
+import com.estudos.nuhome.service.response.HomeUserDetailsResponse
 import com.estudos.nuhome.service.Service
 import com.estudos.nuhome.service.ServiceImpl
-import io.reactivex.Single
 import retrofit2.Call
 
 interface NuHomeRepository {
