+++ app/src/main/java/com/estudos/nuhome/home/domain/NuHomeInteractor.kt
@@ -2,16 +2,16 @@ package com.estudos.nuhome.home.domain
 
 import com.estudos.nuhome.home.data.NuHomeRepository
 import com.estudos.nuhome.home.data.NuHomeRepositoryImpl
-import com.estudos.nuhome.home.data.model.response.HomeUserDetailsResponse
+import com.estudos.nuhome.service.response.HomeUserDetailsResponse
 import retrofit2.Call
-import retrofit2.Callback
-import retrofit2.Response
 
 interface NuHomeInteractor {
+    fun fetchHomeCards(): Call<HomeUserDetailsResponse>
 }
 
-class NuHomeInteractorImpl(val repository: NuHomeRepository = NuHomeRepositoryImpl()) :
-    NuHomeInteractor {
-
-
+class NuHomeInteractorImpl(
+    val repository: NuHomeRepository = NuHomeRepositoryImpl()
+) : NuHomeInteractor {
+    override fun fetchHomeCards(): Call<HomeUserDetailsResponse> =
+        repository.fetchHomeCards()
 }
