+++ app/src/main/res/layout/activity_nu_home.xml
@@ -3,6 +3,12 @@
     xmlns:app="http://schemas.android.com/apk/res-auto"
     xmlns:tools="http://schemas.android.com/tools">
 
+    <data>
+        <variable
+            name="viewmodel"
+            type="com.estudos.nuhome.home.viewmodel.NuHomeViewModel" />
+    </data>
+
     <androidx.core.widget.NestedScrollView
         android:layout_width="match_parent"
         android:layout_height="match_parent"
@@ -60,6 +66,7 @@
 
                         <ImageView
                             android:id="@+id/brandIcon"
+                            android:text="@{viewmodel.infoVO.creditCardIcon}"
                             android:layout_width="@dimen/icon_card_brand_width"
                             android:layout_height="wrap_content"
                             android:layout_marginTop="@dimen/space24"
@@ -72,6 +79,7 @@
                         <TextView
                             android:id="@+id/userName"
                             style="@style/CardHomeText"
+                            android:text="@{viewmodel.infoVO.userName}"
                             android:layout_width="wrap_content"
                             android:layout_height="wrap_content"
                             android:layout_marginStart="@dimen/space16"
@@ -81,7 +89,7 @@
                             app:layout_constraintBottom_toBottomOf="@id/containerCardTopInfo"
                             app:layout_constraintStart_toStartOf="@id/containerCardTopInfo"
                             app:layout_constraintTop_toTopOf="@id/containerCardTopInfo"
-                            tools:text="Nathan" />
+                            tools:text="Mara" />
                     </androidx.constraintlayout.widget.ConstraintLayout>
                 </androidx.cardview.widget.CardView>
             </LinearLayout>
@@ -134,6 +142,7 @@
                         <TextView
                             android:id="@+id/balanceAvailableValue"
                             style="@style/CardHomeText"
+                            android:text="@{viewmodel.infoVO.totalAmoutAvailable}"
                             android:layout_width="wrap_content"
                             android:layout_height="wrap_content"
                             android:layout_marginTop="@dimen/space16"
